+++ src/InfillOp.cc
@@ -1,15 +1,15 @@
 //
-//  InfillOp.m
+//  InfillOp.cc
 //  Mandoline
 //
 //  Created by GM on 11/24/10.
 //  Copyright 2010 Belfry DevWorks. All rights reserved.
 //
 
-#include "InfillOp.h"
-#include "BGL/BGL.h"
-#include "SlicingContext.h"
-#include "CarvedSlice.h"
+#include <BGL.hh>
+#include "InfillOp.hh"
+#include "SlicingContext.hh"
+#include "CarvedSlice.hh"
 
 
 
@@ -25,7 +25,7 @@ void InfillOp::main()
     if ( NULL == context ) return;
     if ( NULL == slice ) return;
 
-    float extrusionWidth = context->standardExtrusionWidth();
+    double extrusionWidth = context->standardExtrusionWidth();
 
     slice->infillMask.infillPathsForRegionWithDensity(context->infillDensity, extrusionWidth, slice->infill);
     slice->state = INFILLED;
